Converted inputlink sagas to listeners
